Add more changes

+++ src/main/java/com/databricks/jdbc/api/impl/IExecutionResult.java
@@ -36,5 +36,5 @@ public interface IExecutionResult {
 
   long getRowCount();
 
-  int getChunkCount();
+  long getChunkCount();
 }
